Move MatrixBatch extensions to separate files
Refactored UserActivity, EventObject, and SessionObject to move their MatrixBatch protocol extensions into dedicated files. This improves code organization and separation of concerns, making the core model files cleaner and easier to maintain.
