+++ Sources/Scout/Core/Log/EventObject.swift
@@ -14,25 +14,3 @@ final class EventObject: SyncableObject, Syncable {
         try batch(in: context, matching: [\.name, \.week])
     }
 }
-
-extension EventObject: MatrixBatch {
-    static func matrix(of batch: [EventObject]) throws(MatrixPropertyError) -> GridMatrix<Int> {
-        guard let name = batch.first?.name else {
-            throw .init("name")
-        }
-        guard let week = batch.first?.week else {
-            throw .init("week")
-        }
-        return Matrix(
-            recordType: "DateIntMatrix",
-            date: week,
-            name: name,
-            cells: parse(of: batch)
-        )
-    }
-
-    static func parse(of batch: [EventObject]) -> [GridCell<Int>] {
-        batch.grouped(by: \.hour).mapValues(\.count).map(Cell.init)
-    }
-}
-
