+++ Sources/Scout/Core/Session/SessionObject+MatrixBatch.swift
@@ -1,0 +1,24 @@
+//
+// Copyright 2025 Mikhail Kasianov
+//
+// Use of this source code is governed by an MIT-style
+// license that can be found in the LICENSE file or at
+// https://opensource.org/licenses/MIT.
+
+extension SessionObject: MatrixBatch {
+    static func matrix(of batch: [SessionObject]) throws(MatrixPropertyError) -> GridMatrix<Int> {
+        guard let week = batch.first?.week else {
+            throw .init("week")
+        }
+        return Matrix(
+            recordType: "DateIntMatrix",
+            date: week,
+            name: "Session",
+            cells: parse(of: batch)
+        )
+    }
+
+    static func parse(of batch: [SessionObject]) -> [GridCell<Int>] {
+        batch.grouped(by: \.date).mapValues(\.count).map(Cell.init)
+    }
+}
